+++ src/Resources/Customer.php
@@ -93,11 +93,11 @@ class Customer extends AbstractUnzerResource
     }
 
     /**
-     * @param string $firstname
+     * @param string|null $firstname
      *
      * @return Customer
      */
-    public function setFirstname($firstname): Customer
+    public function setFirstname(?string $firstname): Customer
     {
         $this->firstname = $firstname;
         return $this;
@@ -112,11 +112,11 @@ class Customer extends AbstractUnzerResource
     }
 
     /**
-     * @param string $lastname
+     * @param string|null $lastname
      *
      * @return Customer
      */
-    public function setLastname($lastname): Customer
+    public function setLastname(?string $lastname): Customer
     {
         $this->lastname = $lastname;
         return $this;
@@ -131,11 +131,11 @@ class Customer extends AbstractUnzerResource
     }
 
     /**
-     * @param string $salutation
+     * @param string|null $salutation
      *
      * @return Customer
      */
-    public function setSalutation($salutation): Customer
+    public function setSalutation(?string $salutation): Customer
     {
         $allowedSalutations = [Salutations::MR, Salutations::MRS, Salutations::UNKNOWN];
         $this->salutation = in_array($salutation, $allowedSalutations, true) ? $salutation : Salutations::UNKNOWN;
@@ -151,11 +151,11 @@ class Customer extends AbstractUnzerResource
     }
 
     /**
-     * @param string $birthday
+     * @param string|null $birthday
      *
      * @return Customer
      */
-    public function setBirthDate($birthday): Customer
+    public function setBirthDate(?string $birthday): Customer
     {
         $this->birthDate = $birthday;
         return $this;
@@ -170,11 +170,11 @@ class Customer extends AbstractUnzerResource
     }
 
     /**
-     * @param string $company
+     * @param string|null $company
      *
      * @return Customer
      */
-    public function setCompany($company): Customer
+    public function setCompany(?string $company): Customer
     {
         $this->company = $company;
         return $this;
@@ -189,11 +189,11 @@ class Customer extends AbstractUnzerResource
     }
 
     /**
-     * @param string $email
+     * @param string|null $email
      *
      * @return Customer
      */
-    public function setEmail($email): Customer
+    public function setEmail(?string $email): Customer
     {
         $this->email = $email;
         return $this;
@@ -208,11 +208,11 @@ class Customer extends AbstractUnzerResource
     }
 
     /**
-     * @param string $phone
+     * @param string|null $phone
      *
      * @return Customer
      */
-    public function setPhone($phone): Customer
+    public function setPhone(?string $phone): Customer
     {
         $this->phone = $phone;
         return $this;
@@ -227,11 +227,11 @@ class Customer extends AbstractUnzerResource
     }
 
     /**
-     * @param string $mobile
+     * @param string|null $mobile
      *
      * @return Customer
      */
-    public function setMobile($mobile): Customer
+    public function setMobile(?string $mobile): Customer
     {
         $this->mobile = $mobile;
         return $this;
@@ -284,11 +284,11 @@ class Customer extends AbstractUnzerResource
     }
 
     /**
-     * @param string $customerId
+     * @param string|null $customerId
      *
      * @return Customer
      */
-    public function setCustomerId($customerId): Customer
+    public function setCustomerId(?string $customerId): Customer
     {
         $this->customerId = $customerId;
         return $this;
@@ -303,11 +303,11 @@ class Customer extends AbstractUnzerResource
     }
 
     /**
-     * @param CompanyInfo $companyInfo
+     * @param CompanyInfo|null $companyInfo
      *
      * @return Customer
      */
-    public function setCompanyInfo(CompanyInfo $companyInfo): Customer
+    public function setCompanyInfo(?CompanyInfo $companyInfo): Customer
     {
         $this->companyInfo = $companyInfo;
         return $this;
@@ -322,7 +322,7 @@ class Customer extends AbstractUnzerResource
     }
 
     /**
-     * @param string $language
+     * @param string|null $language
      *
      * @return Customer
      */
@@ -339,7 +339,7 @@ class Customer extends AbstractUnzerResource
     /**
      * {@inheritDoc}
      */
-    protected function getResourcePath($httpMethod = HttpAdapterInterface::REQUEST_GET): string
+    protected function getResourcePath(string $httpMethod = HttpAdapterInterface::REQUEST_GET): string
     {
         return 'customers';
     }
@@ -359,7 +359,7 @@ class Customer extends AbstractUnzerResource
     /**
      * {@inheritDoc}
      */
-    public function handleResponse(stdClass $response, $method = HttpAdapterInterface::REQUEST_GET): void
+    public function handleResponse(stdClass $response, string $method = HttpAdapterInterface::REQUEST_GET): void
     {
         if (isset($response->companyInfo) && $this->companyInfo === null) {
             $this->companyInfo = new CompanyInfo();