+++ src/Resources/PaymentTypes/Card.php
@@ -66,11 +66,11 @@ class Card extends BasePaymentType
     /**
      * Card constructor.
      *
-     * @param string      $number
-     * @param string      $expiryDate
+     * @param string|null $number
+     * @param string|null $expiryDate
      * @param string|null $email
      */
-    public function __construct($number, $expiryDate, $email = null)
+    public function __construct(?string $number, ?string $expiryDate, string $email = null)
     {
         $this->setNumber($number);
         $this->setExpiryDate($expiryDate);
@@ -88,11 +88,11 @@ class Card extends BasePaymentType
     }
 
     /**
-     * @param string $pan
+     * @param string|null $pan
      *
      * @return Card
      */
-    public function setNumber($pan): Card
+    public function setNumber(?string $pan): Card
     {
         $this->number = $pan;
         return $this;
@@ -107,13 +107,12 @@ class Card extends BasePaymentType
     }
 
     /**
-     * @param string $expiryDate
+     * @param string|null $expiryDate
      *
      * @return Card
      *
-     * @throws RuntimeException
      */
-    public function setExpiryDate($expiryDate): Card
+    public function setExpiryDate(?string $expiryDate): Card
     {
         // Null value is allowed to be able to fetch a card object with nothing but the id set.
         if ($expiryDate === null) {
@@ -138,11 +137,11 @@ class Card extends BasePaymentType
     }
 
     /**
-     * @param string $cvc
+     * @param string|null $cvc
      *
      * @return Card
      */
-    public function setCvc($cvc): Card
+    public function setCvc(?string $cvc): Card
     {
         $this->cvc = $cvc;
         return $this;
@@ -161,7 +160,7 @@ class Card extends BasePaymentType
      *
      * @return Card
      */
-    public function setCardHolder($cardHolder): Card
+    public function setCardHolder(string $cardHolder): Card
     {
         $this->cardHolder = $cardHolder;
         return $this;
@@ -180,7 +179,7 @@ class Card extends BasePaymentType
      *
      * @return Card
      */
-    public function set3ds($card3ds): Card
+    public function set3ds(?bool $card3ds): Card
     {
         $this->card3ds = $card3ds;
         return $this;
@@ -257,7 +256,7 @@ class Card extends BasePaymentType
     /**
      * {@inheritDoc}
      */
-    public function handleResponse(stdClass $response, $method = HttpAdapterInterface::REQUEST_GET): void
+    public function handleResponse(stdClass $response, string $method = HttpAdapterInterface::REQUEST_GET): void
     {
         parent::handleResponse($response, $method);