+++ src/Resources/TransactionTypes/AbstractTransactionType.php
@@ -25,6 +25,7 @@ namespace UnzerSDK\Resources\TransactionTypes;
 use UnzerSDK\Adapter\HttpAdapterInterface;
 use UnzerSDK\Exceptions\UnzerApiException;
 use UnzerSDK\Resources\AbstractUnzerResource;
+use UnzerSDK\Resources\EmbeddedResources\CardTransactionData;
 use UnzerSDK\Resources\EmbeddedResources\RiskData;
 use UnzerSDK\Resources\EmbeddedResources\ShippingData;
 use UnzerSDK\Resources\Payment;
@@ -76,7 +77,7 @@ abstract class AbstractTransactionType extends AbstractUnzerResource
      *
      * @return $this
      */
-    public function setPayment($payment): self
+    public function setPayment(Payment $payment): self
     {
         $this->payment = $payment;
         $this->setParentResource($payment);
@@ -84,9 +85,9 @@ abstract class AbstractTransactionType extends AbstractUnzerResource
     }
 
     /**
-     * Return the Id of the referenced payment object.
+     * Return the ID of the referenced payment object.
      *
-     * @return null|string The Id of the payment object or null if nothing is found.
+     * @return null|string The ID of the payment object or null if nothing is found.
      */
     public function getPaymentId(): ?string
     {
@@ -117,7 +118,7 @@ abstract class AbstractTransactionType extends AbstractUnzerResource
      * @throws UnzerApiException An UnzerApiException is thrown if there is an error returned on API-request.
      * @throws RuntimeException  A RuntimeException is thrown when there is an error while using the SDK.
      */
-    public function handleResponse(stdClass $response, $method = HttpAdapterInterface::REQUEST_GET): void
+    public function handleResponse(stdClass $response, string $method = HttpAdapterInterface::REQUEST_GET): void
     {
         parent::handleResponse($response, $method);
 
@@ -192,6 +193,7 @@ abstract class AbstractTransactionType extends AbstractUnzerResource
 
             $this->handleRiskData($additionalTransactionData);
             $this->handleShipping($additionalTransactionData);
+            $this->handleCardTransactionData($additionalTransactionData);
         }
     }
 
@@ -226,6 +228,23 @@ abstract class AbstractTransactionType extends AbstractUnzerResource
             $shippingObject = $this->getShipping() ?? new ShippingData();
             $shippingObject->handleResponse($shipping);
             $this->setShipping($shippingObject);
+        }
+    }
+
+    /**
+     * Handle CardTransactionData object contained in additional transaction data from API response.
+     *
+     * @param stdClass $additionalTransactionData
+     *
+     * @return void
+     */
+    protected function handleCardTransactionData(stdClass $additionalTransactionData): void
+    {
+        $card = $additionalTransactionData->card ?? null;
+        if ($card !== null) {
+            $cardTransactionData = $this->getCardTransactionData() ?? new CardTransactionData();
+            $cardTransactionData->handleResponse($card);
+            $this->setCardTransactionData($cardTransactionData);
         }
     }
 }