+++ src/Services/IdService.php
@@ -41,7 +41,7 @@ class IdService
      *
      * @throws RuntimeException
      */
-    public static function getResourceIdFromUrl($url, $idString, $onlyLast = false): string
+    public static function getResourceIdFromUrl(string $url, string $idString, bool $onlyLast = false): string
     {
         $matches = [];
         $pattern = '/\/([s|p]{1}-' . $idString . '-[a-z\d]+)\/?' . ($onlyLast ? '$':'') . '/';
@@ -63,7 +63,7 @@ class IdService
      *
      * @throws RuntimeException
      */
-    public static function isPaymentCancellation($url): string
+    public static function isPaymentCancellation(string $url): string
     {
         $pattern = '/\/payments\/[s|p]{1}-pay-[a-z\d]+\/(charges|authorize)\/cancels\/[s|p]{1}-cnl-[a-z\d]+/';
         return preg_match($pattern, $url) === 1;
@@ -78,7 +78,7 @@ class IdService
      *
      * @return string|null
      */
-    public static function getResourceIdOrNullFromUrl($url, $idString, $onlyLast = false): ?string
+    public static function getResourceIdOrNullFromUrl(string $url, string $idString, bool $onlyLast = false): ?string
     {
         try {
             return self::getResourceIdFromUrl($url, $idString, $onlyLast);
@@ -92,22 +92,18 @@ class IdService
      *
      * @return string|null
      */
-    public static function getLastResourceIdFromUrlString($url): ?string
+    public static function getLastResourceIdFromUrlString(string $url): ?string
     {
         return self::getResourceIdOrNullFromUrl($url, '([a-z]{3}|p24)', true);
     }
 
     /**
-     * @param $typeId
+     * @param string $typeId
      *
      * @return string|null
      */
-    public static function getResourceTypeFromIdString(?string $typeId): ?string
+    public static function getResourceTypeFromIdString(string $typeId): ?string
     {
-        if ($typeId === null) {
-            return null;
-        }
-
         $typeIdString = null;
 
         $typeIdParts = [];