+++ test/integration/PaymentTypes/CardTest.php
@@ -26,10 +26,13 @@
 namespace UnzerSDK\test\integration\PaymentTypes;
 
 use UnzerSDK\Constants\ApiResponseCodes;
+use UnzerSDK\Constants\ExemptionType;
 use UnzerSDK\Exceptions\UnzerApiException;
+use UnzerSDK\Resources\EmbeddedResources\CardTransactionData;
 use UnzerSDK\Resources\EmbeddedResources\CardDetails;
 use UnzerSDK\Resources\PaymentTypes\BasePaymentType;
 use UnzerSDK\Resources\PaymentTypes\Card;
+use UnzerSDK\Resources\TransactionTypes\Charge;
 use UnzerSDK\Services\ValueService;
 use UnzerSDK\test\BaseIntegrationTest;
 
@@ -263,6 +266,57 @@ class CardTest extends BaseIntegrationTest
 
         $charge = $card->charge(12.34, 'EUR', 'https://docs.unzer.com');
         $this->assertFalse($charge->isCard3ds());
+    }
+
+    /**
+     * Verfify card transaction can be used with exemptionType
+     *
+     * @test
+     */
+    public function cardTransactionAcceptsExemptionType(): void
+    {
+        $card = $this->createCardObject();
+        /** @var Card $card */
+        $card = $this->unzer->createPaymentType($card);
+        $charge = new Charge(12.34, 'EUR', 'https://docs.unzer.com');
+        $cardTransactionData = (new CardTransactionData())
+            ->setExemptionType(ExemptionType::LOW_VALUE_PAYMENT);
+
+        $charge->setCardTransactionData($cardTransactionData);
+        $this->getUnzerObject()->performCharge($charge, $card);
+
+        // Verify lvp value gets mapped from response
+        $fetchedCharge = $this->unzer->fetchChargeById($charge->getPaymentId(), $charge->getId());
+        $this->assertEquals(ExemptionType::LOW_VALUE_PAYMENT, $fetchedCharge->getCardTransactionData()->getExemptionType());
+    }
+
+    /**
+     * Verify card transaction returns Liability Shift Indicator.
+     *
+     * @test
+     *
+     * @dataProvider cardTransactionReturnsLiabilityIndicatorDP()
+     *
+     * @param mixed $pan
+     */
+    public function cardTransactionReturnsLiabilityIndicator($pan): void
+    {
+        $this->markTestSkipped('Requires a special config for card.');
+        $card = $this->createCardObject()->setNumber($pan)->set3ds(false);
+        /** @var Card $card */
+        $card = $this->unzer->createPaymentType($card);
+        $charge = $card->charge(12.34, 'EUR', 'https://docs.unzer.com');
+
+        // Verify Liability Indicator in response.
+        $this->assertNotNull($charge->getAdditionalTransactionData());
+        $this->assertNotNull($charge->getAdditionalTransactionData()->card->liability);
+
+        // Verify Liability Indicator In payment response.
+        $fetchedPayment = $this->unzer->fetchPayment($charge->getPaymentId());
+        $paymentCharge = $fetchedPayment->getCharge('s-chg-1', true);
+        $this->assertNotNull($paymentCharge->getAdditionalTransactionData()->card->liability);
+
+        $this->getUnzerObject()->fetchCharge($charge);
     }
 
     /**
@@ -668,6 +722,16 @@ class CardTest extends BaseIntegrationTest
         return [
             'invalid string' => ['invalid recurrence Type'],
             'number' => [42],
+        ];
+    }
+
+    public function cardTransactionReturnsLiabilityIndicatorDP()
+    {
+        return [
+            '6799851000000032' => ['6799851000000032'],
+            '5453010000059543' => ['5453010000059543'],
+            '4711100000000000' => ['4711100000000000'],
+            '4012001037461114' => ['4012001037461114']
         ];
     }
 }