+++ test/unit/Constants/PaymentStateTest.php
@@ -34,6 +34,7 @@ class PaymentStateTest extends BasePaymentTest
      * This should verify the mapping of the payment state to the state code.
      *
      * @test
+     *
      * @dataProvider codeToNameDataProvider
      *
      * @param integer $code
@@ -48,6 +49,7 @@ class PaymentStateTest extends BasePaymentTest
      * This should verify the mapping of the payment state to the state code.
      *
      * @test
+     *
      * @dataProvider nameToCodeDataProvider
      *
      * @param integer $code