+++ test/unit/Traits/TraitDummyHasAdditionalTransactionData.php
@@ -26,9 +26,10 @@
 /** @noinspection PhpDocMissingThrowsInspection */
 namespace UnzerSDK\test\unit\Traits;
 
+use UnzerSDK\Resources\AbstractUnzerResource;
 use UnzerSDK\Traits\HasAdditionalTransactionData;
 
-class TraitDummyHasAdditionalTransactionData
+class TraitDummyHasAdditionalTransactionData extends AbstractUnzerResource
 {
     use HasAdditionalTransactionData;
 }